FadeManager 추가
- FadeManager 스크립트 추가
- Fade는 FadeManager에서 관리
