+++ Assets/Scenes/GameScene.unity
@@ -3602,7 +3602,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 532fce04dfd8bda42856fe5a3c78b9e8, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  gameManager: {fileID: 921938699}
+  fadeManager: {fileID: 840480186}
   CameraUI: {fileID: 527049652}
   ExitUI: {fileID: 1080693240}
   fadeImage: {fileID: 1201729923}
@@ -6325,7 +6325,7 @@ MonoBehaviour:
   player: {fileID: 730974879}
   player1: {fileID: 730974888}
   dialogueManager: {fileID: 1638264466}
-  gameManager: {fileID: 921938699}
+  fadeManager: {fileID: 840480186}
 --- !u!81 &286991895
 AudioListener:
   m_ObjectHideFlags: 0
@@ -7478,7 +7478,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 8
+  m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &410108924
 GameObject:
@@ -9389,6 +9389,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   player: {fileID: 730974881}
+  pos: {x: 0, y: 0, z: 0}
   offset: {x: 0, y: 0, z: 0}
 --- !u!1001 &521897455
 PrefabInstance:
@@ -13481,6 +13482,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: cee4200447b99414dbeb4d62911f998f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  pos: {x: 0, y: 0, z: 0}
   stamina: {fileID: 1046106979}
   hp:
   - {fileID: 1554170118}
@@ -13555,6 +13557,7 @@ MonoBehaviour:
   YellowTeddyBear_Sprite: {fileID: 21300000, guid: 62ab146cc215d0e4a9c023859697d1de, type: 3}
   Cake_Sprite: {fileID: 21300000, guid: d522466950a14e34f80ce6df926e5111, type: 3}
   NPC_Sprite: {fileID: 21300000, guid: 4d185be5c15e9ef4c9094c7d531da437, type: 3}
+  canPickup: 0
   d_camera:
     contents:
     - "\uC640!!! \uCE74\uBA54\uB77C\uB2E4! \uCE74\uBA54\uB77C\uB85C! \uCE74\uBA54\uB77C\uB85C....."
@@ -13584,7 +13587,8 @@ MonoBehaviour:
   enemyManager: {fileID: 403567747}
   rend: {fileID: 730974880}
   Player_control: {fileID: 730974887}
-  is_move: 1
+  isMove: 1
+  isPush: 0
   bullet: {fileID: 3386876616857250725, guid: 5634552d58d68674f9c0699ec5e4178b, type: 3}
   bulletPos: {fileID: 658994212}
   fireCooltime: 0.3
@@ -14493,7 +14497,7 @@ MonoBehaviour:
     - {fileID: 21300000, guid: 08ca430d50500a8448f39a19087a625c, type: 3}
     - {fileID: 21300000, guid: 08ca430d50500a8448f39a19087a625c, type: 3}
     - {fileID: 21300000, guid: 08ca430d50500a8448f39a19087a625c, type: 3}
-  gameManager: {fileID: 921938699}
+  fadeManager: {fileID: 840480186}
 --- !u!1 &788612235
 GameObject:
   m_ObjectHideFlags: 0
@@ -15172,6 +15176,52 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 429233718349716679, guid: b36987d2166d3924a847408d854d424a, type: 3}
   m_PrefabInstance: {fileID: 840394519}
   m_PrefabAsset: {fileID: 0}
+--- !u!1 &840480184
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 840480185}
+  - component: {fileID: 840480186}
+  m_Layer: 0
+  m_Name: FadeManager
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &840480185
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 840480184}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: -74.06408, y: 46.491283, z: 0.53735}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 8
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &840480186
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 840480184}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: da712778979fd404191bd8b681963c28, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  fadeImage: {fileID: 1201729923}
+  fadeDuration: 1
 --- !u!1 &841236603
 GameObject:
   m_ObjectHideFlags: 0
@@ -16797,9 +16847,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 5ce68ecfabb1e46439816a25d1e03136, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  dialogueManager: {fileID: 1638264466}
-  fadeImage: {fileID: 1201729923}
-  fadeDuration: 1
+  fadeManager: {fileID: 840480186}
 --- !u!4 &921938700
 Transform:
   m_ObjectHideFlags: 0
@@ -16813,7 +16861,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 7
+  m_RootOrder: 9
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &928426211
 GameObject:
@@ -58024,7 +58072,7 @@ MonoBehaviour:
   player: {fileID: 730974879}
   player1: {fileID: 730974888}
   dialogueManager: {fileID: 1638264466}
-  gameManager: {fileID: 0}
+  fadeManager: {fileID: 840480186}
 --- !u!1 &1235198641
 GameObject:
   m_ObjectHideFlags: 0
