+++ Assets/Script/GameScene/UI/FadeManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: da712778979fd404191bd8b681963c28
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
